feat: add signWithKey and refactor utils
